Allow custom metrics port to be specified

+++ Common/VltkConstants.cs
@@ -4,6 +4,6 @@
     {
         public const string TimestampFormat = "yyyy-MM-ddTHH:mm:ss.fffZ";
 
-        public const int InterpreterMetricServerPort = 19891;
+        public const ushort InterpreterMetricServerPort = 19891;
     }
 }
